projects: aducm_blinky_example: restruct to EXAMPLE build
Move the blinky app to the EXAMPLE-based flow (common/examples/platform)
mirroring eval-adxrs290-pmdz, replacing the app_src layout. Behavior
unchanged.
