+++ projects/aducm_blinky_example/Makefile
@@ -1,9 +1,8 @@
-# Uncomment to use the desired platform
-# PLATFORM = xilinx
-# PLATFORM = altera
-# PLATFORM = aducm3029
+EXAMPLE ?= basic_example
 
 include ../../tools/scripts/generic_variables.mk
+
+include ../../tools/scripts/examples.mk
 
 include src.mk
