+++ projects/aducm_blinky_example/src/common/common_data.h
@@ -1,0 +1,14 @@
+/*****************************************************************************
+ * common_data.h
+ * Defines common data to be used by aducm_blinky_example examples.
+ *****************************************************************************/
+#ifndef __COMMON_DATA_H__
+#define __COMMON_DATA_H__
+
+#include "parameters.h"
+#include "no_os_gpio.h"
+
+extern struct no_os_gpio_init_param blinky_gpio_blue_ip;
+extern struct no_os_gpio_init_param blinky_gpio_green_ip;
+
+#endif /* __COMMON_DATA_H__ */
