+++ projects/aducm_blinky_example/src/platform/aducm3029/platform_src.mk
@@ -1,0 +1,2 @@
+SRC_DIRS += $(PLATFORM_DRIVERS)
+SRC_DIRS += $(INCLUDE)
